requirements.txt
Flask
gunicorn
google-cloud-secret-manager
spacy
pronto
biopython
